added put routes to edit job data
