+++ routes/testRoutes.js
@@ -43,22 +43,4 @@ module.exports = function(app) {
     });
   });
 
-  //--authentic jobs API route
-  app.post("/api/authJobs", function(req, res) {
-    if(check.login(req, res)){
-      return;
-    }
-    var keyword = req.body.jobType;
-    var location = req.body.jobLocation;
-    var queryAuthJobsURL = "https://authenticjobs.com/api/?api_key="+ keys.authenticJobs.key +"&method=aj.jobs.search&category="+ keyword +"&perpage=10&location="+ location +"&format=json";
-    var options = {
-      uri: queryAuthJobsURL,
-      json: true // Automatically parses the JSON string in the response
-    };
-    
-    rp(options).
-    then(function(response){
-      res.json(response)
-    });
-  });
 }
